requirements.txt
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
plotly>=5.20
streamlit>=1.34
ipywidgets>=8.1
pyyaml>=6.0
pytest>=8.0
